+++ build-raml-module-builder.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-#Run this from the root directory of the raml-module-builder source
-
-mvn clean install
-