+++ okapi-registration/managed-deployment/register.sh
@@ -1,9 +1,8 @@
 #!/usr/bin/env bash
 
-module_id=${1}
+deployment_descriptor_filename=${1:-target/DeploymentDescriptor.json}
 okapi_proxy_address=${2:-http://localhost:9130}
 tenant_id=${3:-demo_tenant}
-deployment_descriptor_filename=${4:-target/DeploymentDescriptor.json}
 
 deployment_json=$(cat ./${deployment_descriptor_filename})
 
@@ -18,10 +17,7 @@ curl -w '\n' -D - -s \
      -d @./target/ModuleDescriptor.json  \
      "${okapi_proxy_address}/_/proxy/modules"
 
-activate_json=$(cat ./registration/activate.json)
-activate_json="${activate_json/moduleidhere/$module_id}"
-
 curl -w '\n' -X POST -D - \
      -H "Content-type: application/json" \
-     -d "${activate_json}"  \
+     -d @./target/Activate.json  \
      "${okapi_proxy_address}/_/proxy/tenants/${tenant_id}/modules"