+++ sample-data/import.sh
@@ -11,6 +11,7 @@ contributor_type_storage_address=http://localhost:9130/contributor-types
 instance_format_storage_address=http://localhost:9130/instance-formats
 instance_type_storage_address=http://localhost:9130/instance-types
 classification_type_storage_address=http://localhost:9130/classification-types
+shelf_location_storage_address=http://localhost:9130/shelf-locations
 
 
 for f in ./mtypes/*.json; do
@@ -27,6 +28,14 @@ for f in ./loantypes/*.json; do
          -H "X-Okapi-Tenant: ${tenant}" \
          -d @$f \
          "${loan_type_storage_address}"
+done
+
+for f in ./shelf-locations/*.json; do
+    curl -w '\n' -X POST -D - \
+         -H "Content-type: application/json" \
+         -H "X-Okapi-Tenant: ${tenant}" \
+         -d @$f \
+         "${shelf_location_storage_address}"
 done
 
 for f in ./items/*.json; do