+++ src/test/java/org/folio/rest/api/ShelfLocationsTest.java
@@ -1,0 +1,297 @@
+package org.folio.rest.api;
+
+import io.vertx.core.Handler;
+import io.vertx.core.buffer.Buffer;
+import io.vertx.core.http.HttpClient;
+import io.vertx.core.http.HttpClientRequest;
+import io.vertx.core.http.HttpClientResponse;
+import io.vertx.core.http.HttpMethod;
+import io.vertx.core.json.JsonObject;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import static org.folio.rest.api.StorageTestSuite.itemsUrl;
+import static org.folio.rest.api.StorageTestSuite.loanTypesUrl;
+import static org.folio.rest.api.StorageTestSuite.materialTypesUrl;
+import static org.folio.rest.api.StorageTestSuite.shelfLocationsUrl;
+import org.folio.rest.support.AdditionalHttpStatusCodes;
+import static org.folio.rest.support.HttpResponseMatchers.statusCodeIs;
+import org.folio.rest.support.JsonResponse;
+import org.folio.rest.support.ResponseHandler;
+import org.folio.rest.support.TextResponse;
+import org.folio.rest.support.client.LoanTypesClient;
+import org.folio.rest.support.client.MaterialTypesClient;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import org.junit.Assert;
+import static org.junit.Assert.assertThat;
+import org.junit.Before;
+import org.junit.Test;
+
+
+
+public class ShelfLocationsTest {
+
+  private static final String SUPPORTED_CONTENT_TYPE_JSON_DEF = "application/json";
+  private String canCirculateLoanTypeID;
+  private String journalMaterialTypeID;
+
+  @Before
+  public void beforeEach()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    StorageTestSuite.deleteAll(itemsUrl());
+    StorageTestSuite.deleteAll(shelfLocationsUrl());
+    StorageTestSuite.deleteAll(loanTypesUrl());
+    StorageTestSuite.deleteAll(materialTypesUrl());
+
+    canCirculateLoanTypeID = new LoanTypesClient(
+      new org.folio.rest.support.HttpClient(StorageTestSuite.getVertx()),
+      loanTypesUrl()).create("Can Circulate");
+
+    journalMaterialTypeID = new MaterialTypesClient(
+      new org.folio.rest.support.HttpClient(StorageTestSuite.getVertx()),
+      materialTypesUrl()).create("Journal");
+  }
+
+  @Test
+  public void canCreateShelfLocation()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    JsonResponse response = createShelfLocation("Main Library");
+
+    assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+    assertThat(response.getJson().getString("id"), notNullValue());
+    assertThat(response.getJson().getString("name"), is("Main Library"));
+  }
+
+  @Test
+  public void cannotCreateShelfLocationWithSameName()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    createShelfLocation("Main Library");
+    JsonResponse response = createShelfLocation("Main Library");
+    assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
+  }
+
+  @Test
+  public void cannotCreateShelfLocationWithSameId()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    createShelfLocation(id, "Main Library");
+    JsonResponse response = createShelfLocation(id, "Annex Library");
+    assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
+  }
+
+  @Test
+  public void canGetAShelfLocationById()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    createShelfLocation(id, "Main Library");
+    JsonResponse getResponse = getById(id);
+
+    assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_OK));
+    JsonObject item = getResponse.getJson();
+    assertThat(item.getString("id"), is(id.toString()));
+    assertThat(item.getString("name"), is("Main Library"));
+  }
+
+  @Test
+  public void canUpdateAShelfLocation()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    createShelfLocation(id, "Main Library");
+
+    JsonObject updateRequest = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Annex Library");
+
+    CompletableFuture<TextResponse> updated = new CompletableFuture<>();
+
+    send(shelfLocationsUrl("/" + id.toString()).toString(), HttpMethod.PUT,
+      updateRequest.toString(), SUPPORTED_CONTENT_TYPE_JSON_DEF,
+      ResponseHandler.text(updated));
+
+    TextResponse updateResponse = updated.get(5, TimeUnit.SECONDS);
+
+    assertThat(updateResponse, statusCodeIs(HttpURLConnection.HTTP_NO_CONTENT));
+    assertThat(updateResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    JsonResponse getResponse = getById(id);
+    assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_OK));
+    JsonObject item = getResponse.getJson();
+    assertThat(item.getString("id"), is(id.toString()));
+    assertThat(item.getString("name"), is("Annex Library"));
+  }
+
+  @Test
+  public void canDeleteAShelfLocation()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+
+    createShelfLocation(id, "Main Library");
+
+    CompletableFuture<TextResponse> deleteCompleted = new CompletableFuture<>();
+
+    send(shelfLocationsUrl("/" + id.toString()).toString(), HttpMethod.DELETE, null,
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.text(deleteCompleted));
+
+    TextResponse deleteResponse = deleteCompleted.get(5, TimeUnit.SECONDS);
+    assertThat(deleteResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+  }
+
+  @Test
+  public void cannotDeleteAShelfLocationAssociatedWithAnItem()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID locationId = UUID.randomUUID();
+
+    createShelfLocation(locationId, "Main Library");
+
+    JsonObject item = createItemRequest(locationId.toString());
+    CompletableFuture<JsonResponse> createItemCompleted = new CompletableFuture();
+
+    send(itemsUrl().toString(), HttpMethod.POST, item.toString(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(createItemCompleted));
+
+    JsonResponse createItemResponse = createItemCompleted.get(5, TimeUnit.SECONDS);
+    assertThat(createItemResponse.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    CompletableFuture<TextResponse> deleteCompleted = new CompletableFuture<>();
+
+    send(shelfLocationsUrl(locationId.toString()).toString(),
+      HttpMethod.DELETE, null, SUPPORTED_CONTENT_TYPE_JSON_DEF,
+      ResponseHandler.text(deleteCompleted));
+
+    TextResponse deleteResponse = deleteCompleted.get(5, TimeUnit.SECONDS);
+    assertThat(deleteResponse.getStatusCode(), is(HttpURLConnection.HTTP_BAD_REQUEST));
+  }
+
+  private void send(String url, HttpMethod method, String content,
+                    String contentType, Handler<HttpClientResponse> handler) {
+
+    HttpClient client = StorageTestSuite.getVertx().createHttpClient();
+    HttpClientRequest request;
+
+    if(content == null){
+      content = "";
+    }
+    Buffer buffer = Buffer.buffer(content);
+
+    if (method == HttpMethod.POST) {
+      request = client.postAbs(url);
+    }
+    else if (method == HttpMethod.DELETE) {
+      request = client.deleteAbs(url);
+    }
+    else if (method == HttpMethod.GET) {
+      request = client.getAbs(url);
+    }
+    else {
+      request = client.putAbs(url);
+    }
+    request.exceptionHandler(error -> {
+      Assert.fail(error.getLocalizedMessage());
+    })
+    .handler(handler);
+
+    request.putHeader("Authorization", "test_tenant");
+    request.putHeader("x-okapi-tenant", "test_tenant");
+    request.putHeader("Accept", "application/json,text/plain");
+    request.putHeader("Content-type", contentType);
+    request.end(buffer);
+  }
+
+  private JsonObject createItemRequest(String permanentLocationId) {
+
+    JsonObject item = new JsonObject();
+
+    item.put("title", "abcd");
+    item.put("barcode", "12345");
+    item.put("permanentLoanTypeId", canCirculateLoanTypeID);
+    item.put("materialTypeId", journalMaterialTypeID);
+    item.put("permanentLocationId", permanentLocationId);
+
+    return item;
+  }
+
+  private JsonResponse createShelfLocation(String name)
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    CompletableFuture<JsonResponse> createShelfLocation = new CompletableFuture();
+    String createSLURL = shelfLocationsUrl().toString();
+
+    send(createSLURL, HttpMethod.POST, new JsonObject().put("name", name).toString(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(createShelfLocation));
+
+    return createShelfLocation.get(5, TimeUnit.SECONDS);
+  }
+
+  private JsonResponse createShelfLocation(UUID id, String name)
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    CompletableFuture<JsonResponse> createShelfLocation = new CompletableFuture();
+
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", name);
+
+    send(shelfLocationsUrl().toString(), HttpMethod.POST, request.toString(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(createShelfLocation));
+
+    return createShelfLocation.get(5, TimeUnit.SECONDS);
+  }
+
+  private JsonResponse getById(UUID id)
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    CompletableFuture<JsonResponse> getCompleted = new CompletableFuture<>();
+
+    send(shelfLocationsUrl("/" + id.toString()).toString(), HttpMethod.GET,
+      null, SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(getCompleted));
+
+    return getCompleted.get(5, TimeUnit.SECONDS);
+  }
+
+}
+