+++ NEWS.md
@@ -1,3 +1,10 @@
+## 6.0.0 Unreleased
+
+* Removes `location` property from Item record, and store a UUID for a (permanent and temporary) location record instead
+* Implement `/shelf-locations` endpoint for CRUD of location records
+* Provides `item-storage` 4.0 interface
+* Provides `shelf-locations` 1.0 interface
+
 ## 5.1.1 Unreleased
 
 * Generates Descriptors at build time from templates in ./descriptors (FOLIO-701)